refactor: handle secrets logic in a custom hook
