+++ renderer/src/features/mcp-servers/lib/__tests__/orchestrate-run-local-server.test.tsx
@@ -6,18 +6,12 @@ import {
   convertCreateRequestToFormData,
   prepareUpdateWorkloadData,
 } from '../orchestrate-run-local-server'
-import type { DefinedSecret, PreparedSecret } from '@/common/types/secrets'
 import type {
   SecretsSecretParameter,
   CoreWorkload,
   V1CreateRequest,
   V1ListSecretsResponse,
 } from '@api/types.gen'
-import {
-  getMCPDefinedSecrets,
-  groupMCPDefinedSecrets,
-  saveMCPSecrets,
-} from '@/common/lib/utils'
 
 vi.mock('sonner', async () => {
   const original = await vi.importActual<typeof import('sonner')>('sonner')
@@ -33,105 +27,8 @@ vi.mock('sonner', async () => {
   }
 })
 
-/**
- * Creates reusable mocks for testing the utility functions
- */
-function createTestMocks(options?: {
-  saveSecretImplementation?: () => Promise<{ key: string }>
-}) {
-  return {
-    mockSaveSecret: vi
-      .fn()
-      .mockImplementation(
-        options?.saveSecretImplementation ||
-          (() => Promise.resolve({ key: 'test-key' }))
-      ),
-    mockOnSecretSuccess: vi.fn(),
-    mockOnSecretError: vi.fn(),
-  }
-}
-
 beforeEach(() => {
   vi.clearAllMocks()
-})
-
-describe('groupMCPDefinedSecrets', () => {
-  it('separates new and existing secrets correctly', () => {
-    const secrets: DefinedSecret[] = [
-      {
-        name: 'NEW_SECRET',
-        value: { isFromStore: false, secret: 'new-value' },
-      },
-      {
-        name: 'EXISTING_SECRET',
-        value: { isFromStore: true, secret: 'existing-key' },
-      },
-      {
-        name: 'ANOTHER_NEW_SECRET',
-        value: { isFromStore: false, secret: 'another-new-value' },
-      },
-    ]
-
-    const result = groupMCPDefinedSecrets(secrets)
-
-    expect(result.newSecrets).toHaveLength(2)
-    expect(result.existingSecrets).toHaveLength(1)
-
-    expect(result.newSecrets[0]?.name).toBe('NEW_SECRET')
-    expect(result.newSecrets[1]?.name).toBe('ANOTHER_NEW_SECRET')
-    expect(result.existingSecrets[0]?.name).toBe('EXISTING_SECRET')
-    expect(result.existingSecrets[0]?.value.secret).toBe('existing-key')
-    expect(result.existingSecrets[0]?.value.isFromStore).toBe(true)
-  })
-
-  it('handles empty secrets array', () => {
-    const result = groupMCPDefinedSecrets([])
-
-    expect(result.newSecrets).toHaveLength(0)
-    expect(result.existingSecrets).toHaveLength(0)
-  })
-
-  it('handles all new secrets', () => {
-    const secrets: DefinedSecret[] = [
-      {
-        name: 'SECRET1',
-        value: { isFromStore: false, secret: 'value1' },
-      },
-      {
-        name: 'SECRET2',
-        value: { isFromStore: false, secret: 'value2' },
-      },
-    ]
-
-    const result = groupMCPDefinedSecrets(secrets)
-
-    expect(result.newSecrets).toHaveLength(2)
-    expect(result.existingSecrets).toHaveLength(0)
-  })
-
-  it('handles all existing secrets', () => {
-    const secrets: DefinedSecret[] = [
-      {
-        name: 'SECRET1',
-        value: { isFromStore: true, secret: 'key1' },
-      },
-      {
-        name: 'SECRET2',
-        value: { isFromStore: true, secret: 'key2' },
-      },
-    ]
-
-    const result = groupMCPDefinedSecrets(secrets)
-
-    expect(result.newSecrets).toHaveLength(0)
-    expect(result.existingSecrets).toHaveLength(2)
-    expect(result.existingSecrets[0]?.name).toBe('SECRET1')
-    expect(result.existingSecrets[0]?.value.secret).toBe('key1')
-    expect(result.existingSecrets[0]?.value.isFromStore).toBe(true)
-    expect(result.existingSecrets[1]?.name).toBe('SECRET2')
-    expect(result.existingSecrets[1]?.value.secret).toBe('key2')
-    expect(result.existingSecrets[1]?.value.isFromStore).toBe(true)
-  })
 })
 
 describe('prepareCreateWorkloadData', () => {
@@ -404,189 +301,6 @@ describe('prepareCreateWorkloadData', () => {
   })
 })
 
-describe('saveSecrets', () => {
-  it('saves secrets serially and returns created secrets', async () => {
-    const { mockSaveSecret, mockOnSecretSuccess, mockOnSecretError } =
-      createTestMocks()
-
-    const secrets: PreparedSecret[] = [
-      {
-        secretStoreKey: 'SECRET1',
-        target: 'TARGET1',
-        value: 'value1',
-      },
-      {
-        secretStoreKey: 'SECRET2',
-        target: 'TARGET2',
-        value: 'value2',
-      },
-    ]
-
-    // Mock saveSecret to simulate calling the onSuccess callback
-    mockSaveSecret.mockImplementation(async (request, options) => {
-      const result = { key: request.body.key }
-      if (options?.onSuccess) {
-        options.onSuccess()
-      }
-      return result
-    })
-
-    const result = await saveMCPSecrets(
-      secrets,
-      mockSaveSecret,
-      mockOnSecretSuccess,
-      mockOnSecretError
-    )
-
-    expect(mockSaveSecret).toHaveBeenCalledTimes(2)
-    expect(mockSaveSecret).toHaveBeenNthCalledWith(
-      1,
-      { body: { key: 'SECRET1', value: 'value1' } },
-      expect.any(Object)
-    )
-    expect(mockSaveSecret).toHaveBeenNthCalledWith(
-      2,
-      { body: { key: 'SECRET2', value: 'value2' } },
-      expect.any(Object)
-    )
-
-    expect(mockOnSecretSuccess).toHaveBeenCalledTimes(2)
-    expect(mockOnSecretSuccess).toHaveBeenNthCalledWith(1, 1, 2)
-    expect(mockOnSecretSuccess).toHaveBeenNthCalledWith(2, 2, 2)
-
-    expect(result).toEqual([
-      { name: 'SECRET1', target: 'TARGET1' },
-      { name: 'SECRET2', target: 'TARGET2' },
-    ])
-  })
-
-  it('handles error when saving a secret fails', async () => {
-    const mockError = new Error('Failed to save secret')
-    const { mockSaveSecret, mockOnSecretSuccess, mockOnSecretError } =
-      createTestMocks()
-
-    const secrets: PreparedSecret[] = [
-      {
-        secretStoreKey: 'SECRET1',
-        target: 'TARGET1',
-        value: 'value1',
-      },
-    ]
-
-    // Mock saveSecret to simulate calling the onError callback
-    mockSaveSecret.mockImplementation(async (request, options) => {
-      if (options?.onError) {
-        options.onError(mockError.message, request)
-      }
-      throw mockError
-    })
-
-    await expect(
-      saveMCPSecrets(
-        secrets,
-        mockSaveSecret,
-        mockOnSecretSuccess,
-        mockOnSecretError
-      )
-    ).rejects.toThrow('Failed to save secret')
-
-    expect(mockSaveSecret).toHaveBeenCalledTimes(1)
-    expect(mockOnSecretSuccess).not.toHaveBeenCalled()
-  })
-
-  it('throws error when secret creation returns no key', async () => {
-    const { mockSaveSecret, mockOnSecretSuccess, mockOnSecretError } =
-      createTestMocks()
-
-    const secrets: PreparedSecret[] = [
-      {
-        secretStoreKey: 'SECRET1',
-        target: 'TARGET1',
-        value: 'value1',
-      },
-    ]
-
-    // Mock saveSecret to return empty key and call onSuccess
-    mockSaveSecret.mockImplementation(async (_, options) => {
-      const result = { key: '' }
-      if (options?.onSuccess) {
-        options.onSuccess()
-      }
-      return result
-    })
-
-    await expect(
-      saveMCPSecrets(
-        secrets,
-        mockSaveSecret,
-        mockOnSecretSuccess,
-        mockOnSecretError
-      )
-    ).rejects.toThrow('Failed to create secret for key "SECRET1"')
-  })
-
-  it('handles empty secrets array', async () => {
-    const { mockSaveSecret, mockOnSecretSuccess, mockOnSecretError } =
-      createTestMocks()
-
-    const result = await saveMCPSecrets(
-      [],
-      mockSaveSecret,
-      mockOnSecretSuccess,
-      mockOnSecretError
-    )
-
-    expect(mockSaveSecret).not.toHaveBeenCalled()
-    expect(mockOnSecretSuccess).not.toHaveBeenCalled()
-    expect(result).toEqual([])
-  })
-
-  it('includes delay in non-test environment', async () => {
-    // Save original NODE_ENV
-    const originalEnv = process.env.NODE_ENV
-
-    try {
-      // Set to non-test environment
-      process.env.NODE_ENV = 'development'
-
-      const { mockSaveSecret, mockOnSecretSuccess, mockOnSecretError } =
-        createTestMocks()
-
-      const secrets: PreparedSecret[] = [
-        {
-          secretStoreKey: 'SECRET1',
-          target: 'TARGET1',
-          value: 'value1',
-        },
-      ]
-
-      // Mock saveSecret to simulate calling the onSuccess callback
-      mockSaveSecret.mockImplementation(async (request, options) => {
-        const result = { key: request.body.key }
-        if (options?.onSuccess) {
-          options.onSuccess()
-        }
-        return result
-      })
-
-      const startTime = Date.now()
-      await saveMCPSecrets(
-        secrets,
-        mockSaveSecret,
-        mockOnSecretSuccess,
-        mockOnSecretError
-      )
-      const endTime = Date.now()
-
-      // Should have some delay (at least 100ms based on the implementation)
-      expect(endTime - startTime).toBeGreaterThanOrEqual(100)
-    } finally {
-      // Restore original NODE_ENV
-      process.env.NODE_ENV = originalEnv
-    }
-  })
-})
-
 describe('convertWorkloadToFormData', () => {
   it('converts docker workload to form data', () => {
     const workload: CoreWorkload = {
@@ -696,78 +410,6 @@ describe('convertWorkloadToFormData', () => {
     const result = convertWorkloadToFormData(workload)
 
     expect(result.transport).toBe('stdio')
-  })
-})
-
-describe('getMCPDefinedSecrets', () => {
-  it('filters out secrets with empty names', () => {
-    const secrets: FormSchemaLocalMcp['secrets'] = [
-      {
-        name: 'VALID_SECRET',
-        value: { secret: 'secret-key', isFromStore: false },
-      },
-      {
-        name: '',
-        value: { secret: 'another-key', isFromStore: true },
-      },
-      {
-        name: 'ANOTHER_VALID',
-        value: { secret: 'third-key', isFromStore: false },
-      },
-    ]
-
-    const result = getMCPDefinedSecrets(secrets)
-
-    expect(result).toHaveLength(2)
-    expect(result[0]?.name).toBe('VALID_SECRET')
-    expect(result[1]?.name).toBe('ANOTHER_VALID')
-  })
-
-  it('filters out secrets with empty secret values', () => {
-    const secrets: FormSchemaLocalMcp['secrets'] = [
-      {
-        name: 'VALID_SECRET',
-        value: { secret: 'secret-key', isFromStore: false },
-      },
-      {
-        name: 'INVALID_SECRET',
-        value: { secret: '', isFromStore: false },
-      },
-    ]
-
-    const result = getMCPDefinedSecrets(secrets)
-
-    expect(result).toHaveLength(1)
-    expect(result[0]?.name).toBe('VALID_SECRET')
-  })
-
-  it('preserves isFromStore property', () => {
-    const secrets: FormSchemaLocalMcp['secrets'] = [
-      {
-        name: 'STORE_SECRET',
-        value: { secret: 'key1', isFromStore: true },
-      },
-      {
-        name: 'NEW_SECRET',
-        value: { secret: 'key2', isFromStore: false },
-      },
-      {
-        name: 'UNDEFINED_STORE',
-        value: { secret: 'key3', isFromStore: false },
-      },
-    ]
-
-    const result = getMCPDefinedSecrets(secrets)
-
-    expect(result).toHaveLength(3)
-    expect(result[0]?.value.isFromStore).toBe(true)
-    expect(result[1]?.value.isFromStore).toBe(false)
-    expect(result[2]?.value.isFromStore).toBe(false) // defaults to false
-  })
-
-  it('returns empty array for empty input', () => {
-    const result = getMCPDefinedSecrets([])
-    expect(result).toEqual([])
   })
 })
