+++ renderer/src/features/registry-servers/lib/__tests__/orchestrate-run-registry-server.test.tsx
@@ -1,13 +1,7 @@
 import { it, expect, vi, describe } from 'vitest'
 import type { RegistryImageMetadata } from '@api/types.gen'
 import { prepareCreateWorkloadData } from '../orchestrate-run-registry-server'
-import type { DefinedSecret, PreparedSecret } from '@/common/types/secrets'
 import type { FormSchemaRegistryMcp } from '../form-schema-registry-mcp'
-import {
-  getMCPDefinedSecrets,
-  groupMCPDefinedSecrets,
-  saveMCPSecrets,
-} from '@/common/lib/utils'
 
 const REGISTRY_SERVER: RegistryImageMetadata = {
   name: 'test-server',
@@ -25,147 +19,6 @@ const REGISTRY_SERVER: RegistryImageMetadata = {
 
 beforeEach(() => {
   vi.clearAllMocks()
-})
-
-describe('getMCPDefinedSecrets', () => {
-  it('filters out secrets with empty name or secret value', () => {
-    const secrets: FormSchemaRegistryMcp['secrets'] = [
-      {
-        name: 'GITHUB_API_TOKEN',
-        value: { secret: 'foo-bar', isFromStore: false },
-      },
-      {
-        name: '',
-        value: { secret: 'some-value', isFromStore: false },
-      },
-      {
-        name: 'EMPTY_SECRET',
-        value: { secret: '', isFromStore: false },
-      },
-      {
-        name: 'VALID_SECRET',
-        value: { secret: 'valid-value', isFromStore: true },
-      },
-    ]
-
-    const result = getMCPDefinedSecrets(secrets)
-
-    expect(result).toEqual([
-      {
-        name: 'GITHUB_API_TOKEN',
-        value: { secret: 'foo-bar', isFromStore: false },
-      },
-      {
-        name: 'VALID_SECRET',
-        value: { secret: 'valid-value', isFromStore: true },
-      },
-    ])
-  })
-
-  it('returns empty array when no valid secrets', () => {
-    const secrets: FormSchemaRegistryMcp['secrets'] = [
-      {
-        name: '',
-        value: { secret: 'some-value', isFromStore: false },
-      },
-      {
-        name: 'EMPTY_SECRET',
-        value: { secret: '', isFromStore: false },
-      },
-    ]
-
-    const result = getMCPDefinedSecrets(secrets)
-
-    expect(result).toEqual([])
-  })
-
-  it('handles empty secrets array', () => {
-    const result = getMCPDefinedSecrets([])
-    expect(result).toEqual([])
-  })
-})
-
-describe('groupMCPDefinedSecrets', () => {
-  it('groups secrets into new and existing categories', () => {
-    const secrets: DefinedSecret[] = [
-      {
-        name: 'NEW_SECRET',
-        value: { secret: 'new-value', isFromStore: false },
-      },
-      {
-        name: 'EXISTING_SECRET',
-        value: { secret: 'existing-key', isFromStore: true },
-      },
-      {
-        name: 'ANOTHER_NEW_SECRET',
-        value: { secret: 'another-new-value', isFromStore: false },
-      },
-    ]
-
-    const result = groupMCPDefinedSecrets(secrets)
-
-    expect(result.newSecrets).toEqual([
-      {
-        name: 'NEW_SECRET',
-        value: { secret: 'new-value', isFromStore: false },
-      },
-      {
-        name: 'ANOTHER_NEW_SECRET',
-        value: { secret: 'another-new-value', isFromStore: false },
-      },
-    ])
-
-    expect(result.existingSecrets).toEqual([
-      {
-        name: 'EXISTING_SECRET',
-        value: { secret: 'existing-key', isFromStore: true },
-      },
-    ])
-  })
-
-  it('handles empty secrets array', () => {
-    const result = groupMCPDefinedSecrets([])
-    expect(result).toEqual({
-      newSecrets: [],
-      existingSecrets: [],
-    })
-  })
-
-  it('handles all new secrets', () => {
-    const secrets: DefinedSecret[] = [
-      {
-        name: 'NEW_SECRET_1',
-        value: { secret: 'value-1', isFromStore: false },
-      },
-      {
-        name: 'NEW_SECRET_2',
-        value: { secret: 'value-2', isFromStore: false },
-      },
-    ]
-
-    const result = groupMCPDefinedSecrets(secrets)
-
-    expect(result.newSecrets).toEqual(secrets)
-    expect(result.existingSecrets).toEqual([])
-  })
-
-  it('handles all existing secrets', () => {
-    const secrets: DefinedSecret[] = [
-      {
-        name: 'EXISTING_SECRET_1',
-        value: { secret: 'key-1', isFromStore: true },
-      },
-      {
-        name: 'EXISTING_SECRET_2',
-        value: { secret: 'key-2', isFromStore: true },
-      },
-    ]
-
-    const result = groupMCPDefinedSecrets(secrets)
-
-    expect(result.newSecrets).toEqual([])
-    expect(result.existingSecrets).toEqual(secrets)
-  })
 })
 
 describe('prepareCreateWorkloadData', () => {
@@ -322,133 +175,6 @@ describe('prepareCreateWorkloadData', () => {
     // OPTIONAL_VAR and ANOTHER_OPTIONAL should be omitted
     expect(result.env_vars).not.toContain('OPTIONAL_VAR=')
     expect(result.env_vars).not.toContain('ANOTHER_OPTIONAL=')
-  })
-})
-
-describe('saveSecrets', () => {
-  it('saves secrets serially and calls progress callbacks', async () => {
-    const mockSaveSecret = vi.fn().mockImplementation((body, options) => {
-      // Call the onSuccess callback to simulate successful save
-      options.onSuccess?.()
-      return Promise.resolve({ key: body.body.key + '_KEY' })
-    })
-
-    const onSecretSuccess = vi.fn()
-    const onSecretError = vi.fn()
-
-    const secrets: PreparedSecret[] = [
-      {
-        secretStoreKey: 'SECRET_1',
-        target: 'SECRET_1',
-        value: 'value-1',
-      },
-      {
-        secretStoreKey: 'SECRET_2',
-        target: 'SECRET_2',
-        value: 'value-2',
-      },
-    ]
-
-    const result = await saveMCPSecrets(
-      secrets,
-      mockSaveSecret,
-      onSecretSuccess,
-      onSecretError
-    )
-
-    expect(mockSaveSecret).toHaveBeenCalledTimes(2)
-    expect(mockSaveSecret).toHaveBeenCalledWith(
-      { body: { key: 'SECRET_1', value: 'value-1' } },
-      expect.any(Object)
-    )
-    expect(mockSaveSecret).toHaveBeenCalledWith(
-      { body: { key: 'SECRET_2', value: 'value-2' } },
-      expect.any(Object)
-    )
-
-    expect(onSecretSuccess).toHaveBeenCalledTimes(2)
-    expect(onSecretSuccess).toHaveBeenCalledWith(1, 2)
-    expect(onSecretSuccess).toHaveBeenCalledWith(2, 2)
-
-    expect(onSecretError).not.toHaveBeenCalled()
-
-    expect(result).toEqual([
-      {
-        name: 'SECRET_1_KEY',
-        target: 'SECRET_1',
-      },
-      {
-        name: 'SECRET_2_KEY',
-        target: 'SECRET_2',
-      },
-    ])
-  })
-
-  it('handles empty secrets array', async () => {
-    const mockSaveSecret = vi.fn()
-    const onSecretSuccess = vi.fn()
-    const onSecretError = vi.fn()
-
-    const result = await saveMCPSecrets(
-      [],
-      mockSaveSecret,
-      onSecretSuccess,
-      onSecretError
-    )
-
-    expect(mockSaveSecret).not.toHaveBeenCalled()
-    expect(onSecretSuccess).not.toHaveBeenCalled()
-    expect(onSecretError).not.toHaveBeenCalled()
-    expect(result).toEqual([])
-  })
-
-  it('throws error when saveSecret returns no key', async () => {
-    const mockSaveSecret = vi.fn().mockImplementation((_, options) => {
-      // Call the onSuccess callback to simulate successful save
-      options.onSuccess?.()
-      return Promise.resolve({ key: null })
-    })
-    const onSecretSuccess = vi.fn()
-    const onSecretError = vi.fn()
-
-    const secrets: PreparedSecret[] = [
-      {
-        secretStoreKey: 'SECRET_1',
-        target: 'SECRET_1',
-        value: 'value-1',
-      },
-    ]
-
-    await expect(
-      saveMCPSecrets(secrets, mockSaveSecret, onSecretSuccess, onSecretError)
-    ).rejects.toThrow('Failed to create secret for key "SECRET_1"')
-  })
-
-  it('calls error callback when saveSecret fails', async () => {
-    const mockError = new Error('Save failed')
-    const mockSaveSecret = vi.fn().mockImplementation((body, options) => {
-      // Call the onError callback to simulate failed save
-      options.onError?.(mockError, body)
-      return Promise.reject(mockError)
-    })
-    const onSecretSuccess = vi.fn()
-    const onSecretError = vi.fn()
-
-    const secrets: PreparedSecret[] = [
-      {
-        secretStoreKey: 'SECRET_1',
-        target: 'SECRET_1',
-        value: 'value-1',
-      },
-    ]
-
-    await expect(
-      saveMCPSecrets(secrets, mockSaveSecret, onSecretSuccess, onSecretError)
-    ).rejects.toThrow('Save failed')
-
-    expect(onSecretError).toHaveBeenCalledWith(mockError, {
-      body: { key: 'SECRET_1', value: 'value-1' },
-    })
   })
 
   it('includes network isolation data when enabled', () => {
